[Update] Model
- 모델 폴더 생성 및 기존 모델 이동
    - Currency, Product는 object 혹은 dto 폴더 생성하여 별도 처리 고민 중..
- Product
    - status bool 타입으로 변경
    - toggleStatus 추가
    - setSelf() 삭제
    - setCount() 수량 증가로 변경
- ProductModel
    - getPrice() 삭제
    - isStatusChange() 함수 추가
    - updateCount() 업데이트 target.setCount로 변경
    - 주석 수정
- WalletModel
    - this.TYPES 외부 객체로 변경
    - setInitialCurrencires() 함수 추가
    - this.currencies 및 this.balance init()에서 할당
    - updateCurrency() setCount로 변경
    - 주석 수정
