+++ src/js/model/Currency.js
@@ -4,6 +4,11 @@ export default class Currency {
     this.count = count;
     this.index = index;
   }
+
+  isEmpty() {
+    return this.count === 0;
+  }
+
   setCount(count) {
     this.count += count;
   }
